extract custom expvar wrappers into monitoring pkg
