+++ engine/expvar.go
@@ -3,51 +3,24 @@ package engine
 // TODO: move as mach as possible to cli monitoring
 
 import (
-	"expvar"
 	"log"
-	"strconv"
-	"sync/atomic"
 	"time"
+
+	"github.com/yandex/pandora/monitoring"
 )
 
-type Counter struct {
-	i int64
-}
-
-func (c *Counter) String() string {
-	return strconv.FormatInt(atomic.LoadInt64(&c.i), 10)
-}
-
-func (c *Counter) Add(delta int64) {
-	atomic.AddInt64(&c.i, delta)
-}
-
-func (c *Counter) Set(value int64) {
-	atomic.StoreInt64(&c.i, value)
-}
-
-func (c *Counter) Get() int64 {
-	return atomic.LoadInt64(&c.i)
-}
-
-func NewCounter(name string) *Counter {
-	v := &Counter{}
-	expvar.Publish(name, v)
-	return v
-}
-
 var (
-	evRequests      = NewCounter("engine_Requests")
-	evResponses     = NewCounter("engine_Responses")
-	evUsersStarted  = NewCounter("engine_UsersStarted")
-	evUsersFinished = NewCounter("engine_UsersFinished")
+	evRequests      = monitoring.NewCounter("engine_Requests")
+	evResponses     = monitoring.NewCounter("engine_Responses")
+	evUsersStarted  = monitoring.NewCounter("engine_UsersStarted")
+	evUsersFinished = monitoring.NewCounter("engine_UsersFinished")
 )
 
 func init() {
-	evReqPS := NewCounter("engine_ReqPS")
-	evResPS := NewCounter("engine_ResPS")
-	evActiveUsers := NewCounter("engine_ActiveUsers")
-	evActiveRequests := NewCounter("engine_ActiveRequests")
+	evReqPS := monitoring.NewCounter("engine_ReqPS")
+	evResPS := monitoring.NewCounter("engine_ResPS")
+	evActiveUsers := monitoring.NewCounter("engine_ActiveUsers")
+	evActiveRequests := monitoring.NewCounter("engine_ActiveRequests")
 	requests := evRequests.Get()
 	responses := evResponses.Get()
 	go func() {
